requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-invitations==1.9.3
django-memcached==0.1.2
django-model-utils==4.1.1
django-pandas==0.6.4
django-rest-auth==0.9.5
djangorestframework==3.12.4
environ==1.0
et-xmlfile==1.1.0
idna==2.10
kombu==5.1.0
numpy==1.21.0
oauthlib==3.1.1
openpyxl==3.0.7
pandas==1.2.5
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pymemcache==3.4.4
python-dateutil==2.8.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==1.4.4
